change test names and pagination tests to render pagination
